+++ weblate/utils/tests/test_validators.py
@@ -1,0 +1,56 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright © 2012 - 2017 Michal Čihař <michal@cihar.com>
+#
+# This file is part of Weblate <https://weblate.org/>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+
+from unittest import TestCase
+
+from django.core.exceptions import ValidationError
+
+from weblate.utils.validators import validate_editor
+
+
+class EditorValidatorTest(TestCase):
+    def test_empty(self):
+        self.assertIsNone(validate_editor(''))
+
+    def test_valid(self):
+        self.assertIsNone(
+            validate_editor('editor://open/?file=%(file)s&line=%(line)s')
+        )
+
+    def test_invalid_format(self):
+        self.assertRaises(
+            ValidationError,
+            validate_editor,
+            'editor://open/?file=%(fle)s&line=%(line)s'
+        )
+
+    def test_no_scheme(self):
+        self.assertRaises(
+            ValidationError,
+            validate_editor,
+            './local/url'
+        )
+
+    def test_invalid_scheme(self):
+        self.assertRaises(
+            ValidationError,
+            validate_editor,
+            'javascript:alert(0)'
+        )